MedKLIP-main/requirements.txt
einops==0.4.1
exposure==0.1.1
huggingface-hub==0.6.0
iniconfig==1.1.1
matplotlib==3.5.2
nltk==3.7
# numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
omegaconf==2.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.3.5
Pillow==9.0.1
ruamel-yaml-conda 
ruamel.yaml==0.17.21
scikit-image==0.18.1
scikit-learn==1.0.2
scipy==1.7.3
scispacy==0.5.1
sentencepiece==0.1.97
SimpleITK==2.1.1
spacy==3.4.3
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
torch 
torchmetrics==0.9.3
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.2
tqdm==4.64.0
